==> Analysis/Step1_process_GWAS.R <==
setwd("D:/新建文件夹/A")

A<-vroom("Lung cancer_GCST004748.txt.gz")
colnames(A)
Ncases<-29266 
Ntotal<-85716
v<-Ncases/(Ntotal)
TotalNeff<-4*v*(1-v)*(Ntotal)
A$Neff <- 4 / ((2 * A$EAF * (1 - A$EAF)) * (A$SE^2))
A$Neff<-ifelse(A$Neff > 1.1*TotalNeff, 1.1*TotalNeff, A$Neff)
A$Neff<-ifelse(A$Neff < 0.5*TotalNeff, 0.5*TotalNeff, A$Neff)
A$Neff <- round(A$Neff, 2)
A$Neff<-as.numeric(A$Neff)
A<-subset(A,SE>0.00001)
A <- subset(A, abs(BETA) <= 3)
A$MAF<-eaf2maf(A$EAF)
A<-subset(A,MAF>0.01)
A<-na.omit(A)
A<-dplyr::select(A,c(SNP,CHR,POS,effect_allele,other_allele,BETA,SE,P,EAF,Neff))
colnames(A)<-c("SNP","CHR","POS","effect_allele","other_allele","BETA","SE","P","EAF","N")
write_tsv(A,file = "LC_GCST004748.txt.gz")
gc()

A<-vroom("Endometrial cancer_GCST006464.txt.gz")
colnames(A)
Ncases<-12906 
Ntotal<-121885
v<-Ncases/(Ntotal)
TotalNeff<-4*v*(1-v)*(Ntotal)
A$Neff <- 4 / ((2 * A$EAF * (1 - A$EAF)) * (A$SE^2))
A$Neff<-ifelse(A$Neff > 1.1*TotalNeff, 1.1*TotalNeff, A$Neff)
A$Neff<-ifelse(A$Neff < 0.5*TotalNeff, 0.5*TotalNeff, A$Neff)
A$Neff <- round(A$Neff, 2)
A$Neff<-as.numeric(A$Neff)
A<-subset(A,SE>0.00001)
A <- subset(A, abs(BETA) <= 3)
A$MAF<-eaf2maf(A$EAF)
A<-subset(A,MAF>0.01)
A<-na.omit(A)
A<-dplyr::select(A,c(SNP,CHR,POS,effect_allele,other_allele,BETA,SE,P,EAF,Neff))
colnames(A)<-c("SNP","CHR","POS","effect_allele","other_allele","BETA","SE","P","EAF","N")
write_tsv(A,file = "EC_GCST006464.txt.gz")
gc()

A<-vroom("Breast cancer_GCST010098.txt.gz")
colnames(A)
Ncases<-133384 
Ntotal<-247173
v<-Ncases/(Ntotal)
TotalNeff<-4*v*(1-v)*(Ntotal)
A$Neff <- 4 / ((2 * A$EAF * (1 - A$EAF)) * (A$SE^2))
A$Neff<-ifelse(A$Neff > 1.1*TotalNeff, 1.1*TotalNeff, A$Neff)
A$Neff<-ifelse(A$Neff < 0.5*TotalNeff, 0.5*TotalNeff, A$Neff)
A$Neff <- round(A$Neff, 2)
A$Neff<-as.numeric(A$Neff)
A<-subset(A,SE>0.00001)
A <- subset(A, abs(BETA) <= 3)
A$MAF<-eaf2maf(A$EAF)
A<-subset(A,MAF>0.01)
A<-na.omit(A)
A<-dplyr::select(A,c(SNP,CHR,POS,effect_allele,other_allele,BETA,SE,P,EAF,Neff))
colnames(A)<-c("SNP","CHR","POS","effect_allele","other_allele","BETA","SE","P","EAF","N")
write_tsv(A,file = "BRCA_GCST010098.txt.gz")
gc()

A<-vroom("Prostate cancer_GCST90274714.txt.gz")
colnames(A)
Ncases<-122188 
Ntotal<-726828
v<-Ncases/(Ntotal)
TotalNeff<-4*v*(1-v)*(Ntotal)
A$Neff <- 4 / ((2 * A$EAF * (1 - A$EAF)) * (A$SE^2))
A$Neff<-ifelse(A$Neff > 1.1*TotalNeff, 1.1*TotalNeff, A$Neff)
A$Neff<-ifelse(A$Neff < 0.5*TotalNeff, 0.5*TotalNeff, A$Neff)
A$Neff <- round(A$Neff, 2)
A$Neff<-as.numeric(A$Neff)
A<-subset(A,SE>0.00001)
A <- subset(A, abs(BETA) <= 3)
A$MAF<-eaf2maf(A$EAF)
A<-subset(A,MAF>0.01)
A<-na.omit(A)
A<-dplyr::select(A,c(SNP,CHR,POS,effect_allele,other_allele,BETA,SE,P,EAF,Neff))
colnames(A)<-c("SNP","CHR","POS","effect_allele","other_allele","BETA","SE","P","EAF","N")
write_tsv(A,file = "PRCA_GCST90274714.txt.gz")
gc()

A<-vroom("ovarian cancer_GCST004462.txt.gz")
colnames(A)
Ncases<-16924 
Ntotal<-85426
v<-Ncases/(Ntotal)
TotalNeff<-4*v*(1-v)*(Ntotal)
A$Neff <- 4 / ((2 * A$EAF * (1 - A$EAF)) * (A$SE^2))
A$Neff<-ifelse(A$Neff > 1.1*TotalNeff, 1.1*TotalNeff, A$Neff)
A$Neff<-ifelse(A$Neff < 0.5*TotalNeff, 0.5*TotalNeff, A$Neff)
A$Neff <- round(A$Neff, 2)
A$Neff<-as.numeric(A$Neff)
A<-subset(A,SE>0.00001)
A <- subset(A, abs(BETA) <= 3)
A$MAF<-eaf2maf(A$EAF)
A<-subset(A,MAF>0.01)
A<-na.omit(A)
A<-dplyr::select(A,c(SNP,CHR,POS,effect_allele,other_allele,BETA,SE,P,EAF,Neff))
colnames(A)<-c("SNP","CHR","POS","effect_allele","other_allele","BETA","SE","P","EAF","N")
write_tsv(A,file = "OC_GCST004462.txt.gz")
gc()


A<-vroom("CRC_GCST90255675.txt.gz")
colnames(A)
Ncases<-78473
Ntotal<-107143
v<-Ncases/(Ntotal)
TotalNeff<-4*v*(1-v)*(Ntotal)
TotalNeff
A$Neff <- 4 / ((2 * A$EAF * (1 - A$EAF)) * (A$SE^2))
A$Neff<-ifelse(A$Neff > 1.1*TotalNeff, 1.1*TotalNeff, A$Neff)
A$Neff<-ifelse(A$Neff < 0.5*TotalNeff, 0.5*TotalNeff, A$Neff)
A$Neff <- round(A$Neff, 2)
A$Neff<-as.numeric(A$Neff)
A<-subset(A,SE>0.00001)
A <- subset(A, abs(BETA) <= 3)
A$MAF<-eaf2maf(A$EAF)
A<-subset(A,MAF>0.01)
A<-na.omit(A)
A<-dplyr::select(A,c(SNP,CHR,POS,effect_allele,other_allele,BETA,SE,P,EAF,Neff))
colnames(A)<-c("SNP","CHR","POS","effect_allele","other_allele","BETA","SE","P","EAF","N")
write_tsv(A,file = "CRC_GCST90255675.txt.gz")
gc()

A<-vroom("TC_GCST90399736.txt.gz")
colnames(A)
Ncases<-6015
Ntotal<-1339769
v<-Ncases/(Ntotal)
TotalNeff<-4*v*(1-v)*(Ntotal)
TotalNeff
A$Neff <- 4 / ((2 * A$EAF * (1 - A$EAF)) * (A$SE^2))
A$Neff<-ifelse(A$Neff > 1.1*TotalNeff, 1.1*TotalNeff, A$Neff)
A$Neff<-ifelse(A$Neff < 0.5*TotalNeff, 0.5*TotalNeff, A$Neff)
A$Neff <- round(A$Neff, 2)
A$Neff<-as.numeric(A$Neff)
A<-subset(A,SE>0.00001)
A <- subset(A, abs(BETA) <= 3)
A$MAF<-eaf2maf(A$EAF)
A<-subset(A,MAF>0.01)
A<-na.omit(A)
A<-dplyr::select(A,c(SNP,CHR,POS,effect_allele,other_allele,BETA,SE,P,EAF,Neff))
colnames(A)<-c("SNP","CHR","POS","effect_allele","other_allele","BETA","SE","P","EAF","N")
write_tsv(A,file = "THCA_GCST90399736.txt.gz")
gc()


A<-vroom("Esophageal adenocarcinoma_GCST003739.txt.gz")
B<-vroom("reference.1000G.maf.0.005.txt.gz")
B<-dplyr::select(B,c(SNP,MAF))
A<-left_join(A,B,by = "SNP")

colnames(A)
Ncases<-4112 
Ntotal<-21271
v<-Ncases/(Ntotal)
TotalNeff<-4*v*(1-v)*(Ntotal)
A$Neff <- 4 / ((2 * A$MAF * (1 - A$MAF)) * (A$SE^2))
A$Neff<-ifelse(A$Neff > 1.1*TotalNeff, 1.1*TotalNeff, A$Neff)
A$Neff<-ifelse(A$Neff < 0.5*TotalNeff, 0.5*TotalNeff, A$Neff)
A$Neff <- round(A$Neff, 2)
A$Neff<-as.numeric(A$Neff)
A<-subset(A,SE>0.00001)
A <- subset(A, abs(BETA) <= 3)
A<-subset(A,MAF>0.01)
A<-na.omit(A)
A<-dplyr::select(A,c(SNP,CHR,POS,effect_allele,other_allele,BETA,SE,P,MAF,Neff))
colnames(A)<-c("SNP","CHR","POS","effect_allele","other_allele","BETA","SE","P","EAF","N")
write_tsv(A,file = "ESCA_GCST003739.txt.gz")
gc()


A<-vroom("Kidney cancer_GCST90320057.txt.gz")
colnames(A)
A$v<-A$Ncase/(A$N)
A$Neff<-4*A$v*(1-A$v)*(A$N)
Ncases<-25890 
Ntotal<-769475
v<-Ncases/(Ntotal)
TotalNeff<-4*v*(1-v)*(Ntotal)
TotalNeff
A$Neff<-ifelse(A$Neff > 1.1*TotalNeff, 1.1*TotalNeff, A$Neff)
A$Neff<-ifelse(A$Neff < 0.5*TotalNeff, 0.5*TotalNeff, A$Neff)
A$Neff <- round(A$Neff, 2)
A$Neff<-as.numeric(A$Neff)
A<-subset(A,SE>0.00001)
A <- subset(A, abs(BETA) <= 3)
A$MAF<-eaf2maf(A$EAF)
A<-subset(A,MAF>0.01)
A<-na.omit(A)
A<-dplyr::select(A,c(SNP,CHR,POS,effect_allele,other_allele,BETA,SE,P,EAF,Neff))
colnames(A)<-c("SNP","CHR","POS","effect_allele","other_allele","BETA","SE","P","EAF","N")
write_tsv(A,file = "RCC_GCST90320057.txt.gz")
gc()
